requirements.txt
# Third party libraries
aio-pika==9.4.3
aiomisc==17.5.26
asyncpg==0.29.0
croniter==3.0.3
dishka==1.3.0
fastapi-slim==0.115.0
faststream==0.5.25
granian[reload]==1.6.0
httpx==0.27.2
orjson==3.10.7
prometheus-client==0.21.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
starlette==0.38.6
typer==0.12.5
typing-extensions==4.12.2
uvloop==0.20.0
